Move Enums into their own file
